Preliminary getting variables from the debugger
