+++ src/client/datascience/interactive-window/interactiveWindow.ts
@@ -102,7 +102,7 @@ export class InteractiveWindow extends InteractiveBase implements IInteractiveWi
         @inject(IWorkspaceService) workspaceService: IWorkspaceService,
         @inject(IInteractiveWindowProvider) private interactiveWindowProvider: IInteractiveWindowProvider,
         @inject(IDataViewerProvider) dataExplorerProvider: IDataViewerProvider,
-        @inject(IJupyterVariables) jupyterVariables: IJupyterVariables,
+        @inject(IJupyterVariables) @named(Identifiers.ALL_VARIABLES) jupyterVariables: IJupyterVariables,
         @inject(IJupyterDebugger) jupyterDebugger: IJupyterDebugger,
         @inject(IDataScienceErrorHandler) errorHandler: IDataScienceErrorHandler,
         @inject(IPersistentStateFactory) private readonly stateFactory: IPersistentStateFactory,
